digitme: implement handler for BufEnter BufLeave
When enter/leave a buff, the client should be notified in order to keep
a correct session history about the file edited. More info about the
buff is needed, like project the file belongs to.

Also, it's about time to benchmark the ping event because I want to
finalise this feature and not worry about it anymore.